Redrum
I built it one part at a time and it didn't cost me a dime.

+++ pkg/crows/job/condition.go
@@ -54,7 +54,7 @@ func (c *Condition) Parse(rawSearch []byte) Result {
 
 		if val, ok := hack["total_results"]; ok {
 			return Result{
-				Count: val.(int),
+				Count: int(val.(float64)),
 				When:  time.Now(),
 			}
 		}
